shingleindex/requirements.txt
annotated-types==0.7.0
datasketch==1.6.5
nats-py==2.9.0
numpy==2.2.1
pydantic==2.10.4
pydantic_core==2.27.2
redis==5.2.1
scipy==1.15.0
typing_extensions==4.12.2
